+++ database/migrations/2023_12_22_162130_create_product_units_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('product_units', function (Blueprint $table) {
             $table->id();
             $table->foreignId("unit_id");
-            $table->decimal("price", 15, 2);
+            $table->integer("price");
             $table->foreignId("product_id");
         });
     }
